--- src/LinkedInAuth.js
@@ -1,4 +1,0 @@
-import React from 'react'
-import SocialAuth from './SocialAuth'
-
-export default props => <SocialAuth {...props} provider="linkedin" />
